+++ packages/nakadachi-adapter-remix/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rym-lib/nakadachi-adapter-remix
 
+## 1.2.7
+
+### Patch Changes
+
+- feat(query-module): support SQL expression objects in function-based rules #51
+- Updated dependencies
+  - @rym-lib/nakadachi@1.2.7
+
 ## 1.2.6
 
 ### Patch Changes