+++ packages/query-module/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rym-lib/query-module
 
+## 1.2.7
+
+### Patch Changes
+
+- feat(query-module): support SQL expression objects in function-based rules #51
+- Updated dependencies
+  - @rym-lib/exception@1.2.7
+
 ## 1.2.6
 
 ### Patch Changes