+++ packages/remix-helper/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rym-lib/remix-helper
 
+## 1.2.7
+
+### Patch Changes
+
+- feat(query-module): support SQL expression objects in function-based rules #51
+- Updated dependencies
+  - @rym-lib/utilities@1.2.7
+
 ## 1.2.6
 
 ### Patch Changes